requirements/ingest-outlook.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements/ingest-outlook.in
#
certifi==2023.7.22
    # via
    #   -c requirements/base.txt
    #   -c requirements/constraints.in
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.0
    # via
    #   -c requirements/base.txt
    #   requests
cryptography==41.0.4
    # via
    #   msal
    #   pyjwt
idna==3.4
    # via
    #   -c requirements/base.txt
    #   requests
msal==1.24.1
    # via
    #   -r requirements/ingest-outlook.in
    #   office365-rest-python-client
office365-rest-python-client==2.4.2
    # via
    #   -c requirements/constraints.in
    #   -r requirements/ingest-outlook.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.8.0
    # via msal
pytz==2023.3.post1
    # via office365-rest-python-client
requests==2.31.0
    # via
    #   -c requirements/base.txt
    #   msal
    #   office365-rest-python-client
urllib3==1.26.17
    # via
    #   -c requirements/base.txt
    #   -c requirements/constraints.in
    #   requests
